requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-adminactions==1.8.1
django-grappelli==2.13.3
django-grappelli-extras==1.2.1
django-import-export==2.0.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
